# Improve vectorization of Enumerable.Min/Max