- Expand it to all vectorizable comparable types with `Enumerable.Min<T>`/`Max<T>` (byte, sbyte, short, ushort, uint, ulong, nuint, nint) beyond the int/long already supported
- Combine Min/Max into a single method using a static abstract interface with generic specialization to differentiate the operations
- Use `Vector128<T>` and `Vector256<T>` instead of `Vector<T>`
- Process any remaining elements as part of a single final vector rather than falling back to a scalar loop
- Improve test coverage

#### // Int32Tests (int was previously vectorized)

| Method |         Toolchain | Count |      Mean | Ratio |
|------- |------------------ |------ |----------:|------:|
|    Max | \main\corerun.exe |     1 |  2.111 ns |  1.00 |
|    Max |   \pr\corerun.exe |     1 |  1.978 ns |  0.94 |
|        |                   |       |           |       |
|    Max | \main\corerun.exe |     4 |  4.719 ns |  1.00 |
|    Max |   \pr\corerun.exe |     4 |  3.186 ns |  0.67 |
|        |                   |       |           |       |
|    Max | \main\corerun.exe |     7 |  7.473 ns |  1.00 |
|    Max |   \pr\corerun.exe |     7 |  3.372 ns |  0.45 |
|        |                   |       |           |       |
|    Max | \main\corerun.exe |    23 | 11.092 ns |  1.00 |
|    Max |   \pr\corerun.exe |    23 |  4.940 ns |  0.45 |
|        |                   |       |           |       |
|    Max | \main\corerun.exe |  1024 | 99.851 ns |  1.00 |
|    Max |   \pr\corerun.exe |  1024 | 43.150 ns |  0.43 |

#### // ByteTests (bytes weren't previously vectorized)

| Method |         Toolchain | Count |          Mean | Ratio |
|------- |------------------ |------ |--------------:|------:|
|    Max | \main\corerun.exe |     1 |     16.055 ns |  1.00 |
|    Max |   \pr\corerun.exe |     1 |      3.556 ns |  0.22 |
|        |                   |       |               |       |
|    Max | \main\corerun.exe |    16 |     73.951 ns |  1.00 |
|    Max |   \pr\corerun.exe |    16 |     15.251 ns |  0.21 |
|        |                   |       |               |       |
|    Max | \main\corerun.exe |    28 |    126.087 ns |  1.00 |
|    Max |   \pr\corerun.exe |    28 |     15.454 ns |  0.12 |
|        |                   |       |               |       |
|    Max | \main\corerun.exe |    92 |    387.411 ns |  1.00 |
|    Max |   \pr\corerun.exe |    92 |     29.735 ns |  0.08 |
|        |                   |       |               |       |
|    Max | \main\corerun.exe |  4096 | 16,702.786 ns | 1.000 |
|    Max |   \pr\corerun.exe |  4096 |     69.616 ns | 0.004 |

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System;
using System.Linq;

public partial class Program
{
    static void Main(string[] args) => BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
}

public class Int32Tests
{
    [Params(1, 4, 7, 23, 1024)]
    public int Count { get; set; }

    private int[] _values;

    [GlobalSetup]
    public void Setup()
    {
        var r = new Random(42);
        _values = Enumerable.Range(0, Count).Select(_ => r.Next()).ToArray();
    }

    [Benchmark]
    public int Max() => _values.Max();
}

public class ByteTests
{
    [Params(1, 16, 28, 92, 4096)]
    public int Count { get; set; }

    private byte[] _values;

    [GlobalSetup]
    public void Setup()
    {
        var r = new Random(42);
        _values = Enumerable.Range(0, Count).Select(_ => (byte)r.Next()).ToArray();
    }

    [Benchmark]
    public byte Max() => _values.Max();
}
```