@EgorBo 2022-09-25
I suspect that LINQ Max/Min are not IEEE754 compatible for floating points due to usage of `Vector_.Max/Min` which are lowered to single `vmaxps` instruction.

So e.g. `-0.0f` can be considered as Max over `+0.0f`

cc @tannergooding