@stephentoub 2022-09-25
> I suspect that LINQ Max/Min are not IEEE754 compatible for floating points due to usage of Vector_.Max/Min

This PR isn't using vevtor for floating point.