@EgorBo 2022-09-25
I suspect this is not right for floating points as Vector128/256.Max/Min are not 

> > I suspect that LINQ Max/Min are not IEEE754 compatible for floating points due to usage of Vector_.Max/Min
> 
> This PR isn't using vector for floating point.

May I ask why then? Since it won't break the current behavior