@stephentoub 2022-09-25
> May I ask why then? Since it won't break the current behavior

This pr is focused on the MaxInteger method, which doesn't have to be concerned with nans.