@stephentoub 2022-09-25
>> May I ask why then? Since it won't break the current behavior

> This pr is focused on the MaxInteger method, which doesn't have to be concerned with nans.

@EgorBo, FYI, I just tried vectorizing Max's use (still handling the search for the first non-NaN sequentially and then vectorizing the rest), and it breaks existing tests around NaN.