@EgorBo 2022-09-25
> > > May I ask why then? Since it won't break the current behavior
> 
> > This pr is focused on the MaxInteger method, which doesn't have to be concerned with nans.
> 
> @EgorBo, FYI, I just tried vectorizing Max's use (still handling the search for the first non-NaN sequentially and then vectorizing the rest), and it breaks existing tests around NaN.

Ah I see, thanks for trying it! I assume we need to raise an issue for this behavior since it's not consistent, will wait @tannergooding's feedback whether if it's expected or there is already an issue (I failed to find one)

I assume the IEEE754 Max simd function might look like this:
```csharp
static Vector128<float> Max(Vector128<float> v1, Vector128<float> v2) =>
    (Vector128.Max(v1, v2) & Vector128.Max(v1, v2)) | Sse.CompareUnordered(v1, v2);
```
and just `Vector128.Max` on arm. but it will break the current contract