@tannergooding 2022-09-26
> I assume we need to raise an issue for this behavior since it's not consistent, will wait @tannergooding's feedback whether if it's expected or there is already an issue (I failed to find one)

This is a "downside" of the `Vector<T>` APIs being exposed back in 2014 and without much consideration for IEEE 754 at the time. Instead, most of the code was simplify codified only for performance instead.

Realistically, we should have the `Max`/`Min` APIs on the `Vector<T>` types match the same APIs on `float`/`double`/`Half`. That will technically be a breaking change (although not one most people would notice) and would also negatively impact perf. -- The same should really also be true for the LINQ APIs, but they also follow a "different" behavior.

In an "ideal" world, I'd take the break here and consider exposing a `FastMax`/`FastMin` for the case where you want effectively `(a > b) ? a : b` and `(a < b) ? a : b`, as this is what `maxps`/`minps` effectively due (they return `b` if either input is `nan` or if both results are zero, regardless of sign).